pyproject.toml
[tool.poetry]
name = "upsert_agent"
version = "0.1.0"
description = ""
authors = ["ssabrut <michael.gunawan2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
python-dotenv = "^1.0.1"
crewai = {extras = ["tools"], version = "^0.30.11"}
langchain-groq = "^0.1.5"

[tool.poetry.scripts]
upsert_agent = "src.upsert_agent.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
